+++ frontend/src/components/SearchBar.vue
@@ -25,7 +25,7 @@
           :disabled="!searchQuery"
           @click="search"
         >
-          <v-icon>mdi-magnify</v-icon>
+          <v-icon>fa-solid fa-magnifying-glass</v-icon>
         </v-btn>
       </template>
     </v-autocomplete>